Refactor cli into modules
